Retroactive changes can only apply to last change

+++ app/src/main/java/projects/medicationtracker/SimpleClasses/Medication.java
@@ -24,7 +24,7 @@ public class Medication implements Cloneable, Parcelable {
     private String alias;
     private LocalDateTime[] times;
     private long medId;
-    private long medFrequency;
+    private int medFrequency;
     private int medDosage;
     private boolean active;
     private Medication parent = null;
@@ -44,7 +44,7 @@ public class Medication implements Cloneable, Parcelable {
      * @param medAlias  An alias for the Medication to appear in notifications
      */
     public Medication(String thisMed, String patient, String units, LocalDateTime[] time,
-                      LocalDateTime firstDate, long id, long frequency, int dosage, String medAlias) {
+                      LocalDateTime firstDate, long id, int frequency, int dosage, String medAlias) {
         medName = thisMed;
         patientName = patient;
         medDosageUnits = units;
@@ -74,7 +74,7 @@ public class Medication implements Cloneable, Parcelable {
         patientName = in.readString();
         alias = in.readString();
         medId = in.readLong();
-        medFrequency = in.readLong();
+        medFrequency = in.readInt();
         medDosage = in.readInt();
     }
 
@@ -113,7 +113,7 @@ public class Medication implements Cloneable, Parcelable {
      *
      * @return Medication frequency
      */
-    public long getFrequency() {
+    public int getFrequency() {
         return medFrequency;
     }
 
@@ -221,7 +221,7 @@ public class Medication implements Cloneable, Parcelable {
      *
      * @param medFrequency Frequency of the medication
      */
-    public void setFrequency(long medFrequency) {
+    public void setFrequency(int medFrequency) {
         this.medFrequency = medFrequency;
     }
